[WIP] hw: ide: add `READ LOG EXT` support
Motivation behind this is so that we can show
`Trusted Computing Feature Set` availability to a host on log probe.
Currently only used for DMTF SPDM over the trusted send/recv commands,
as defined in the DMTF DSP0286.

For example, the linux kernel will probe the log page and look for the
`Trusted Computing Support` bit before allowing trusted `send/recv`commands.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>